Add a patch to fix issues with pvmonitors in EPICS base
If you have a waveform-type array (aai, waveform, etc.) and have caputlog
loaded and initialised, then the IOC will no longer send out pvmonitors
on changes (though it still sends on CA monitors). This seems to be due
to a bug in PVAccessCPP about caching of PV channels: when we hit the
caPutLog callback code on a put, we call get_array_info from the record
rset code, which overwrites the requested field with the internal array
buffer. This means that we modify the original pchan in memory which
later causes a field comparison in the db event code (db_post_events) to
fail.

This patch is a minimal fix to caPutLog to get rid of the problem. While
the actual issue seems to be a PVAccessCPP one (it seems that the latest
version of PVXS does not have this problem), it seems simpler to fix it
here instead of digging deeply into code that ultimately will be no
longer in use in a few years.

Note that while the issue was first observed in the context of PVMonitors,
it does seem to be more generic than that and is thus not likely only an
issue with PVAccessCPP; see
https://epics.anl.gov/tech-talk/2024/msg00481.php for another example of
a related issue that this change fixes.

Co-authored-by: Lucas A. M. Magalhães <lucmaga@gmail.com>
Co-authored-by: Gabriel Fedel <gabriel.fedel@ess.eu>
Co-authored-by: Mateusz Nabywaniec <mateusz.nabywaniec@ess.eu>