Fix array comparisons
For the implementation of the JSON logger, the array comparison was left out.
As such, when checking if the value had changed, we would only compare the
first of the values from an array. Thus a change from
```
[1, 2, 3] -> [1, 3, 4, 5, 6, 7]
```
would be listed as not changing, which is not the correct behaviour.