Remove constant default_burst_timeout
This seems to be a bit awkward to get to compile for older c++ versions;
for c++11, you can use
```c
static constexpr double default_burst_timeout = 5.0;
```
but for older c++ (with no `constexpr`) this is not valid; moreover, both of
```c
const double default_burst_timeout = 5.0;
static const double default_burst_timeout = 5.0;
```
fail for different reasons. Given that this timeout is only used in one place
this seems like the simplest fix.